requirements.txt
netmiko>=3.4.0,<4.0.0
alive-progress>=2.1.0,<3.0.0
ntc_templates>=3.0.0,<4.0.0
ruamel.yaml>=0.17.19,<1.0.0
requests >=2.27.1,<3.0.0